requirements.txt
msal>=1.24.0
requests>=2.31.0
pyyaml>=6.0
python-dateutil>=2.8.2
